Programação da ficha : faltando a ordenação da lista ,quando houver a remoção de uma serie , de um exercicio e suas series por completo e a movimentação da serie ! metodos para reordenar na TreinoDAO,ControleTreino e GUIFichaTreino

+++ trunk/WorkOutSystem/src/workoutsystem/model/Treino.java
@@ -10,7 +10,7 @@ import net.sf.oval.constraint.NotEmpty;
 
 public class Treino implements Serializable{
 	
-	private long codigoTreino;
+	private int codigoTreino;
 	@NotEmpty(message = "O treino deve ter um nome")
 	@Length(max = 10 ,message = "Escolha um nome com até 10 caracteres")
 	private String nomeTreino;
@@ -25,7 +25,7 @@ public class Treino implements Serializable{
 		return nomeTreino;
 	}
 	
-	public void setCodigo(long codigoTreino) {
+	public void setCodigo(int codigoTreino) {
 		this.codigoTreino = codigoTreino;
 	}
 	public void setNome(String nomeTreino) {
